ARCH-1: extract set dispatch arm into commands::set::run (iter-225)
